# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReportScope": {
        "enum": [
          "MEMBER",
          "TEAM",
          "MULTI_TEAM"
        ],
        "title": "ReportScope",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WeeklyGenerateRequest": {
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/ReportScope"
          },
          "subjectUserId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjectuserid"
          },
          "teamId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teamid"
          },
          "teamIds": {
            "items": {
              "type": "string"
            },
            "title": "Teamids",
            "type": "array"
          },
          "weekStart": {
            "format": "date",
            "title": "Weekstart",
            "type": "string"
          }
        },
        "required": [
          "scope",
          "weekStart"
        ],
        "title": "WeeklyGenerateRequest",
        "type": "object"
      },
      "WeeklyReportRead": {
        "properties": {
          "confirmedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmedat"
          },
          "contentJson": {
            "additionalProperties": true,
            "title": "Contentjson",
            "type": "object"
          },
          "generationMetadata": {
            "additionalProperties": true,
            "title": "Generationmetadata",
            "type": "object"
          },
          "generationSource": {
            "title": "Generationsource",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inputRecordIds": {
            "items": {
              "type": "string"
            },
            "title": "Inputrecordids",
            "type": "array"
          },
          "missingContributors": {
            "items": {
              "type": "string"
            },
            "title": "Missingcontributors",
            "type": "array"
          },
          "scope": {
            "$ref": "#/components/schemas/ReportScope"
          },
          "status": {
            "$ref": "#/components/schemas/WeeklyReportStatus"
          },
          "subjectUserId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjectuserid"
          },
          "supersedesId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedesid"
          },
          "teamId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teamid"
          },
          "teamIds": {
            "items": {
              "type": "string"
            },
            "title": "Teamids",
            "type": "array"
          },
          "templateId": {
            "title": "Templateid",
            "type": "string"
          },
          "templateVersion": {
            "title": "Templateversion",
            "type": "integer"
          },
          "weekEnd": {
            "format": "date",
            "title": "Weekend",
            "type": "string"
          },
          "weekStart": {
            "format": "date",
            "title": "Weekstart",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scope",
          "subjectUserId",
          "teamId",
          "weekStart",
          "weekEnd",
          "contentJson",
          "missingContributors",
          "inputRecordIds",
          "generationSource",
          "generationMetadata",
          "status",
          "templateId",
          "templateVersion",
          "supersedesId",
          "confirmedAt"
        ],
        "title": "WeeklyReportRead",
        "type": "object"
      },
      "WeeklyReportStatus": {
        "enum": [
          "DRAFT",
          "GENERATED",
          "EDITED",
          "CONFIRMED"
        ],
        "title": "WeeklyReportStatus",
        "type": "string"
      },
      "WeeklyUpdateRequest": {
        "properties": {
          "contentJson": {
            "additionalProperties": true,
            "title": "Contentjson",
            "type": "object"
          }
        },
        "required": [
          "contentJson"
        ],
        "title": "WeeklyUpdateRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Project Assistant API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/teams/{team_id}/templates/active": {
      "get": {
        "operationId": "get_active_template_api_v1_teams__team_id__templates_active_get",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ReportScope"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Active Template Api V1 Teams  Team Id  Templates Active Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Active Template",
        "tags": [
          "weekly-reports"
        ]
      }
    },
    "/api/v1/weekly-reports/generate": {
      "post": {
        "operationId": "generate_weekly_report_api_v1_weekly_reports_generate_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeeklyGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyReportRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Weekly Report",
        "tags": [
          "weekly-reports"
        ]
      }
    },
    "/api/v1/weekly-reports/{report_id}": {
      "get": {
        "operationId": "get_weekly_report_api_v1_weekly_reports__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "teamId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Teamid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyReportRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Weekly Report",
        "tags": [
          "weekly-reports"
        ]
      },
      "put": {
        "operationId": "update_weekly_report_api_v1_weekly_reports__report_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "teamId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Teamid"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeeklyUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyReportRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Weekly Report",
        "tags": [
          "weekly-reports"
        ]
      }
    },
    "/api/v1/weekly-reports/{report_id}/confirm": {
      "post": {
        "operationId": "confirm_weekly_report_api_v1_weekly_reports__report_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "teamId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Teamid"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyReportRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Weekly Report",
        "tags": [
          "weekly-reports"
        ]
      }
    },
...
    "/api/v1/weekly-reports/{report_id}/revisions": {
      "post": {
        "operationId": "create_weekly_revision_api_v1_weekly_reports__report_id__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "teamId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Teamid"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyReportRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Weekly Revision",
        "tags": [
          "weekly-reports"
        ]
      }
    },
...
  }
}